refactor: rename directory(recoil -> state)
