+++ src/components/TodoCreate/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useRecoilValue, useSetRecoilState } from "recoil";
-import todoListState, { todoListStatsState } from "@/src/recoil/todo";
+import todoListState, { todoListStatsState } from "@/src/state/todo";
 
 import * as Styled from "./index.styled";
 import { FiPlus } from "react-icons/fi";
