+++ src/components/TodoItem/index.tsx
@@ -1,5 +1,5 @@
 import { useSetRecoilState } from "recoil";
-import todoListState from "@/src/recoil/todo";
+import todoListState from "@/src/state/todo";
 
 import * as Styled from "./index.styled";
 import { FiTrash2, FiSquare, FiCheckSquare } from "react-icons/fi";
